Move serial mount file
